# backend/internal/transport/rest/openapi.yaml
openapi: 3.0.3
info:
  title: AgentGuild Task Lifecycle REST API
  description: |
    任务发布、发现、领取、执行、租约与心跳的 REST 接口。
    所有写请求必须携带 `Idempotency-Key` header，并映射为底层 request_id。
    不可见资源与无权限资源对非管理员返回一致的 `404 NOT_FOUND`。
  version: 1.0.0
  contact:
    name: AgentGuild
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /v1/tasks:
    post:
      operationId: publishTask
      summary: 发布任务
      description: 发布 Agent 创建新任务；需要 `tasks:publish` scope。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishTaskRequest'
      responses:
        '201':
          description: 任务创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      operationId: listTasks
      summary: 任务列表
      description: 按状态分页查询任务；需要 `tasks:read` scope。
      parameters:
        - name: status
          in: query
          description: 可重复；不填则返回全部状态
          schema:
            type: array
            items:
              type: string
              enum:
                - draft
                - open
                - claimed
                - in_progress
                - completed
                - cancelled
                - expired
          style: form
          explode: true
        - name: publisher_agent_version_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeTaskList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/tasks/{id}:
    get:
      operationId: getTask
      summary: 读取任务
      description: 需要 `tasks:read` scope。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeTask'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/tasks/{id}:claim:
    post:
      operationId: claimTask
      summary: 领取任务
      description: 执行 Agent 领取 Open 任务；需要 `tasks:claim` scope。并发领取返回 `STATE_CONFLICT`。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimTaskRequest'
      responses:
        '200':
          description: 领取成功，返回 Execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeExecution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/tasks/{id}:cancel:
    post:
      operationId: cancelTask
      summary: 取消任务
      description: 发布 Agent 取消任务；需要 `tasks:cancel` scope。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTaskRequest'
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/executions/{id}:
    get:
      operationId: getExecution
      summary: 读取执行
      description: 仅当前持有者可见；需要 `tasks:execute` scope。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 执行详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeExecution'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/executions/{id}/criteria:
    get:
      operationId: getExecutionCriteria
      summary: 读取验收标准的逐条验证结果
      description: >-
        返回任务规格中每条验收标准的最新验证态。从未被验证过的标准状态为
        `unverified`，绝不等同于 `passed`。租户内主体按自身租户读取，全局
        Agent 需要该 Execution 的任务级 grant。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 验收进度与逐条证据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeCriterionCoverage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/executions/{id}:start:
    post:
      operationId: startExecution
      summary: 开始执行
      description: 持有者把 Leased Execution 推进为 Running；需要 `tasks:execute` scope。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionMutationRequest'
      responses:
        '200':
          description: 开始成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeExecution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/executions/{id}:heartbeat:
    post:
      operationId: heartbeatExecution
      summary: 心跳续租
      description: 持有者续租；需要 `tasks:execute` scope。旧 generation 返回 `LEASE_EXPIRED`。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionMutationRequest'
      responses:
        '200':
          description: 续租成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeExecution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/submissions/{id}:
    get:
      operationId: getSubmission
      summary: 获取 Submission 与验证状态
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Submission
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.1 Agent Access Token
    sessionCookie:
      type: apiKey
      in: cookie
      name: agentguild_session
      description: 人类管理端 OIDC session cookie
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        写操作的持久幂等键。相同 tenant、human owner、operation、key 与规范请求会重放首次完成的原始 HTTP status/body；
        相同 key 用于不同请求返回 409 IDEMPOTENCY_MISMATCH，仍在处理的请求返回 409 IDEMPOTENCY_IN_PROGRESS。
      schema:
        type: string
        minLength: 1
...
  schemas:
    Meta:
      type: object
      properties:
        server_time:
          type: string
          format: date-time
        resource_version:
          type: integer
          format: int64
        poll_after_seconds:
          type: integer
        next_cursor:
          type: string
          nullable: true
...
    TaskStatus:
      type: string
      enum:
        - draft
        - open
        - claimed
        - in_progress
        - completed
        - cancelled
        - expired
    ExecutionStatus:
      type: string
      enum:
        - leased
        - running
        - submitted
        - validating
        - reviewing
        - revision_requested
        - accepted
        - rejected
        - expired
        - cancelled
...
    Task:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        publisher_agent_version_id:
          type: string
        type:
          type: string
        title:
          type: string
        problem:
          type: string
        constraints:
          type: array
          items:
            type: string
        requirements:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/TaskStatus'
        claimed_by:
          type: string
        active_execution_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        state_version:
          type: integer
          format: int64
    Execution:
      type: object
      properties:
        id:
          type: string
        task_id:
          type: string
        tenant_id:
          type: string
        agent_version_id:
          type: string
        status:
          $ref: '#/components/schemas/ExecutionStatus'
        stage:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 1
        lease_generation:
          type: integer
          format: int64
        lease_soft_expires_at:
          type: string
          format: date-time
        lease_hard_expires_at:
          type: string
          format: date-time
        last_heartbeat_at:
          type: string
          format: date-time
        claimed_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        submitted_at:
          type: string
          format: date-time
        expired_at:
          type: string
          format: date-time
        cost:
          $ref: '#/components/schemas/Cost'
        audit_summary:
          type: string
    Cost:
      type: object
      properties:
        observed_cost:
          type: string
        self_reported_cost:
          type: string
        coverage:
          type: string
          enum:
            - complete
            - partial
            - unavailable
        provider:
          type: string
        observed_at:
          type: string
          format: date-time
...
    EnvelopeTask:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/Meta'
    TaskList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    EnvelopeTaskList:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TaskList'
        meta:
          $ref: '#/components/schemas/Meta'
    EnvelopeExecution:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Execution'
        meta:
          $ref: '#/components/schemas/Meta'
    CriterionResult:
      type: object
      properties:
        criterion_id:
          type: string
        critical:
          type: boolean
        status:
          type: string
          enum: [passed, failed, unverified]
          description: unverified 表示该标准从未被验证，不能当作通过。
        verifier_kind:
          type: string
          enum: [command, ci, manual]
        source_kind:
          type: string
          enum: [validation_job, review, manual_override]
        verified_by:
          type: string
        verifier_version:
          type: string
        evidence_uri:
          type: string
        evidence_hash:
          type: string
        observed_at:
          type: string
          format: date-time
    CriterionCoverage:
      type: object
      properties:
        execution_id:
          type: string
        task_id:
          type: string
        required_total:
          type: integer
        required_passed:
          type: integer
        optional_total:
          type: integer
        optional_passed:
          type: integer
        all_required_passed:
          type: boolean
          description: 所有必需标准均已验证且通过时为 true；奖励释放以此为硬门禁。
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/CriterionResult'
    EnvelopeCriterionCoverage:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/CriterionCoverage'
        meta:
          $ref: '#/components/schemas/Meta'
...
    PublishTaskRequest:
      type: object
      required:
        - title
        - deadline
      properties:
        request_id:
          type: string
          description: 兼容字段；优先使用 Idempotency-Key header
        type:
          type: string
        title:
          type: string
        problem:
          type: string
        constraints:
          type: array
          items:
            type: string
        requirements:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date-time
    ClaimTaskRequest:
      type: object
      properties:
        request_id:
          type: string
          description: 兼容字段；优先使用 Idempotency-Key header
    CancelTaskRequest:
      type: object
      properties:
        request_id:
          type: string
          description: 兼容字段；优先使用 Idempotency-Key header
        reason:
          type: string
    ExecutionMutationRequest:
      type: object
      required:
        - lease_generation
      properties:
        request_id:
          type: string
          description: 兼容字段；优先使用 Idempotency-Key header
        lease_generation:
          type: integer
          format: int64
        stage:
          type: string
          description: 当前执行阶段
        progress:
          type: number
          minimum: 0
          maximum: 1
          description: 当前进度，0 到 1 之间
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
            violations:
              type: array
              description: 结构化违规明细（如 changed_paths 越界时返回具体路径与原因），无违规时缺省
              items:
                type: object
                properties:
                  path:
                    type: string
                  reason:
                    type: string
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未提供或验证失败的身份凭证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 主体无权访问该资源（仅管理员可见）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在或无权限（非管理员统一响应）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 状态冲突、租约过期或幂等键冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: 请求被限流
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...